Conf for doxygen
MediaWiki core is using Doxygen to generates it is documentation. It is
reasonably fast.

Provide a Doxyfile generated with version 1.8.8 with all comments and
default settings stripped out.

Assumes Graphviz is installed by setting HAVE_DOT = YES.

The documentation is generated by running doxygen at the root of this
repository, the doc is made available at doc/html/index.html

The setup is the same I have made for oojs/ui:
    https://gerrit.wikimedia.org/r/#/c/168490/

Bug: 73530
Change-Id: I3377a370bad65307968dd3f90433277daffebabb