test: Disable phpunit "strictCoverage" option
Replace "strict" (deprecated in PHPUnit 4.5, though we're not
using that yet) with separate options so that we can exclude
strictCoverage, which is currently preventing phpunit-coverage
from being able to collect any data for wikimedia/cdb (everything
ends up marked as "risky" and ignored).

The @covers annotation still make it strict about what to
include in the coverage report.

The behaviour of strictCoverage has never been conventional at
Wikimedia. The reason it is included now is because we upgraded
from PHPUnit 3.7 to 4.4, and the strictCoverage option was not
previously part of "strict".

Change-Id: If893df6cd5e1007b8856942b8b64a43234d6e314