Avoid reading past the last key in Cdb\Reader\PHP::nextkey()
Iterating over keys using the pure-PHP reader results in an exception, since
the method does not know where to stop and will eventually attempt to read past
the last key in the file. The test did not catch this because it only attempted
to read as many keys as it had written.

Fix this by iterating over the file index to find the offset of the first hash
table. Since the hash tables go after the records (see http://cr.yp.to/cdb/cdb.txt)
the position of the first hash table is where the records stop.

Do this the first time ::firstkey() is invoked (rather than in the constructor)
to avoid doing pointless work in the cases where the reader is only used to
look up specific keys rather than iterate over all keys.

This patch also changes the name of the `hashTable` property to `index`. The
first 2048 bytes are not a hash table but a set of pointers to hash tables.

Bug: T119364
Change-Id: If7ea5f12659aae44c5da3a1db1b31c2328495b4c