travis-ci: Fix operation ordering
The change in f8dfa29 was largely a copy-n-paste from similar changes
needed by composer-merge-plugin. I neglected to adjust for the order of
operations used in this travis.yml script. Specifically the
"before_script" hook runs after "install" and not before.
"before_install" is the proper hook to use in this case.

Bug: T137507
Change-Id: Iab07ef86be268d39a8dac8c0705f79d4f6b9dbfe