Add return types and make Reader and Writer signatures more consistent
This specifically does not add native parameter type hints to public
methods like get(), set() and exists() as we indeed currently permit
non-strings to (accidentally or otherwise) make it to these methods in
existing consumer code. Doing so would be a semver-major breaking
change.

Instead, only add return types, and improve code, docs,
and tests for the existing parameter types.

* Fix warnings in Writer/PHP::set() from internal write() and strlen()
  functions still receiving non-string.

* Fix inconsistency around Reader::get(). The Reader\PHP class
  recognised "mixed key" and contained strval() casting. Apply the
  same to the Reader interface and the Reader\DBA implementation.

Change-Id: I39bfcb4fcda830649ff9de367f05a41a4a54209e