Inline 1-use methods for performance/readability
I suggest this for a series of reasons:

1. Performance. Every function scope is expensive in PHP. This code
is executed hundreds of thousands of times. This adds up fast.

2. This allows us to merge two write() calls together.

3. Readabilty becomes better, I would argue. Names like "addbegin"
and "addend" were not really meaningful to begin with. Inlining them
allows to change the execution order slightly and group stuff together
that belongs together. Both $this->hplist and $this->numentries are
only needed much later in finish().

This patch alone makes an actual difference. For example, the
MessageIndexTest in the Translate extension improves from 1.7s to
1.4s on my machine.

Change-Id: I94ce310384a807b8ff177a5c80826d5e77f184b0